+++ src/components/Body.jsx
@@ -1,64 +1,10 @@
 import React from "react";
 import { Container, Card, Icon, Header } from "semantic-ui-react";
 import "./Body.css";
+import datas from '../data/body_cardata';
 
 const Body = () => {
-  const datas = [
-    {
-      header: "Test Automation",
-      description: `Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-                    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s,
-                    when an unknown printer took a galley of type and scrambled it to make a 
-                    type specimen book. It has survived not only five centuries`,
-      icon: "code icon",
-      color: "red",
-    },
-    {
-      header: "Cloud Computing",
-      description: `Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-                            Lorem Ipsum has been the industry's standard dummy text ever since the 1500s,
-                            when an unknown printer took a galley of type and scrambled it to make a 
-                            type specimen book. It has survived not only five centuries`,
-      icon: "cloud icon",
-      color: "orange",
-    },
-    {
-      header: "Analytics",
-      description: `Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-                    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s,
-                    when an unknown printer took a galley of type and scrambled it to make a 
-                    type specimen book. It has survived not only five centuries`,
-      icon: "chart bar icon",
-      color: "yellow",
-    },
-    {
-      header: "AI/ML",
-      description: `Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-                    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s,
-                    when an unknown printer took a galley of type and scrambled it to make a 
-                    type specimen book. It has survived not only five centuries`,
-      icon: "desktop icon",
-      color: "olive",
-    },
-    {
-      header: "DevOPs",
-      description: `Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-                    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s,
-                    when an unknown printer took a galley of type and scrambled it to make a 
-                    type specimen book. It has survived not only five centuries`,
-      icon: "server icon",
-      color: "green",
-    },
-    {
-      header: "Mobile Development",
-      description: `,Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-                    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s,
-                    when an unknown printer took a galley of type and scrambled it to make a 
-                    type specimen book. It has survived not only five centuries`,
-      icon: "mobile alternate icon",
-      color: "teal",
-    },
-  ];
+
 
   return (
     <div>
